requirements.txt
Django==5.1.2
django-environ==0.11.2
djangorestframework==3.15.2
psycopg2-binary==2.9.10
drf-spectacular==0.27.2
